Update misleading and outdated audio capture docs (#116)
* Update misleading and outdated audio capture docs

The old and never properly working way of audio merging (with ffmpeg) should have been removed from the docs with the addition of the proper way of exporting. Otherwise users get confused. Also we dont save .wav files anymore. Docs, like implementation should be kept as minimal as possible. The overhead is not space on the page but maintaining and making sure all content stays up-do-date.

* fix typos and formatting.